Fix moved session hook imports
