+++ packages/talon-chat/src/session/useSessionResources.ts
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import { fetchResourceFromGateway } from "./resourceLoader";
 import type { GatewayClientLike } from "./TalonSessionTypes";
 import type { ResourceViewModel } from "../lib/resourceUris";
-import { useResourcePane } from "./useResourcePane";
+import { useResourcePane } from "./hooks/useResourcePane";
 
 type UseSessionResourcesOptions = {
   agent: string;
